+++ src/main/java/rocks/zipcode/atm/bank/AccountData.java
@@ -10,9 +10,9 @@ public final class AccountData {
     private final String email;
     private String type;
 
-    private final int balance;
+    private final float balance;
 
-    AccountData(int id, String name, String email, int balance) {
+    AccountData(int id, String name, String email, Float balance) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -31,7 +31,7 @@ public final class AccountData {
         return email;
     }
 
-    public int getBalance() {
+    public float getBalance() {
         return balance;
     }
